Porder item/managements

+++ backend/smart-logistics/src/main/java/com/douzon/smartlogistics/domain/item/dao/ItemDao.java
@@ -31,7 +31,7 @@ public class ItemDao {
 
     @Transactional
     public void modify(Integer itemCode, ItemModifyDto itemModifyDto) {
-        Long retrieveItemCode = retrieveItem(itemCode);
+        Integer retrieveItemCode = retrieveItem(itemCode);
 
         itemMapper.modify(retrieveItemCode, itemModifyDto);
     }
@@ -39,12 +39,12 @@ public class ItemDao {
 
     @Transactional
     public void delete(Integer itemCode) {
-        Long retrieveItemCode = retrieveItem(itemCode);
+        Integer retrieveItemCode = retrieveItem(itemCode);
 
         itemMapper.delete(retrieveItemCode);
     }
 
-    private Long retrieveItem(Integer itemCode) {
+    private Integer retrieveItem(Integer itemCode) {
         return itemMapper.retrieve(itemCode).orElseThrow(() -> {
             throw new NoSuchElementException("해당 아이템은 존재하지 않습니다.");
         }).getItemCode();
